+++ gui/GuiManager.h
@@ -16,7 +16,7 @@ class GuiManager : public InputHandler
 {
 private:
     LinkedList<GuiFrame*> frameStack;
-    int currentCount = 0;
+    unsigned long currentCount = 0;
 public:
     GuiManager();
     ~GuiManager();
@@ -27,7 +27,7 @@ public:
 
     void handleArrow(int code) override;
     void handleInput(int code) override;
-    void handleAnimationFrame(int count) override;
+    void handleAnimationFrame(unsigned long count) override;
 };
